requirements.txt
crewai
crewai_tools
acp-sdk
load_dotenv
uv
nest-asyncio
colorama
smolagents
ollama
ollama_chat
fastacp